+++ src/index.html
@@ -9,20 +9,10 @@
 </head>
 
 <body>
-  <!-- <div class="panel">
-Col: <input type="text" class="panel-add__cell" placeholder="Add cell" value="">
-Row: <input type="text" class="panel-add__column" placeholder="Add column" value="">
-    <button class="panel-btn">Add</button>
-  </div>
-  <br><hr><br> -->
-
-  <div class="table">
-
-  </div>
 
   <div class="table">
     <div class="table-content">
-      <div class="table-content__item-header">A</div>
+      <div class="table-content__item header">A</div>
       <div class="table-content__item">text</div>
       <div class="table-content__item">text</div>
       <div class="table-content__item">text</div>
@@ -33,7 +23,7 @@ Row: <input type="text" class="panel-add__column" placeholder="Add column" value
     </div>
 
     <div class="table-content">
-      <div class="table-content__item-header">B</div>
+      <div class="table-content__item header">B</div>
       <div class="table-content__item">text</div>
       <div class="table-content__item">text</div>
       <div class="table-content__item">text</div>
@@ -44,7 +34,7 @@ Row: <input type="text" class="panel-add__column" placeholder="Add column" value
     </div>
 
     <div class="table-content">
-      <div class="table-content__item-header">C</div>
+      <div class="table-content__item header">C</div>
       <div class="table-content__item">text</div>
       <div class="table-content__item">text</div>
       <div class="table-content__item">text</div>
