+++ src/main.js
@@ -1,2 +1,2 @@
 import "./scss/main.scss";
-import "./js/table.js";
+import "./js/main.js";
