+++ src/scss/main.scss
@@ -44,7 +44,7 @@
       border-top: none;
     }
 
-    &__item-header {
+    .header {
       text-align: center;
       min-width: 50px;
       max-width: 100%;
